refactor(infra): keep home paths in one module
