+++ src/infra/home-dir.ts
@@ -86,6 +86,10 @@ export function resolveRequiredOsHomeDir(
   return resolveOsHomeDir(env, homedir) ?? path.resolve(process.cwd());
 }
 
+export function resolveHomeDir(): string | undefined {
+  return resolveEffectiveHomeDir();
+}
+
 /** Expands leading `~`, `~/`, or `~\` with the effective home when one is known. */
 export function expandHomePrefix(
   input: string,
@@ -163,4 +167,53 @@ export function resolveOsHomeRelativePath(
     return path.resolve(expanded);
   }
   return path.resolve(trimmed);
+}
+
+function resolveHomeDisplayPrefix(): { home: string; prefix: string } | undefined {
+  const home = resolveHomeDir();
+  if (!home) {
+    return undefined;
+  }
+  const explicitHome = process.env.OPENCLAW_HOME?.trim();
+  if (explicitHome) {
+    return { home, prefix: "$OPENCLAW_HOME" };
+  }
+  return { home, prefix: "~" };
+}
+
+export function shortenHomePath(input: string): string {
+  if (!input) {
+    return input;
+  }
+  const display = resolveHomeDisplayPrefix();
+  if (!display) {
+    return input;
+  }
+  const { home, prefix } = display;
+  if (input === home) {
+    return prefix;
+  }
+  if (input.startsWith(`${home}/`) || input.startsWith(`${home}\\`)) {
+    return `${prefix}${input.slice(home.length)}`;
+  }
+  return input;
+}
+
+export function shortenHomeInString(input: string): string {
+  if (!input) {
+    return input;
+  }
+  const display = resolveHomeDisplayPrefix();
+  if (!display) {
+    return input;
+  }
+  return input.split(display.home).join(display.prefix);
+}
+
+export function displayPath(input: string): string {
+  return shortenHomePath(input);
+}
+
+export function displayString(input: string): string {
+  return shortenHomeInString(input);
 }
